507 attempted contracts - 156 successful - 15,600 dryrun validations all in around 2 minutes
