+++ Makefile
@@ -44,19 +44,29 @@ test-unit:
 
 build-and-test: build lint test-unit
 
+A = 3
+P = 13
+Q = 13
+M = 10
+N = 10
+DEMO_PROCS=auto
+graviton-demo-run:
+	A=$(A) P=$(P) Q=$(Q) M=$(M) N=$(N) pytest -sv -n $(DEMO_PROCS) --durations=10 tests/demo/logicsigs_test.py::test_many_factorizer_games
+
 # set NUM_PROCS = auto when the following issue has been fixed https://github.com/algorand/pyteal/issues/199
 NUM_PROCS = 1 
-integration-test: pip-integration build
+integration-setup: pip-integration build
 	black --check tests $(ALLPY)
 	flake8 tests $(ALLPY)
 	mypy tests
+
+integration-run:
 	pytest -n $(NUM_PROCS) --durations=10 -sv tests/integration
+
+integration-test: integration-setup integration-run graviton-demo-run
 
 # Extras:
 coverage:
 	pytest --cov-report html --cov=pyteal
-
-graviton-demo:
-	pytest -n $(NUM_PROCS) --durations=10 -svv tests/integration/logicsigs_test.py::test_factorizer_game
 
 .PHONY: build docs demo
